backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Facebook Ads Library Backend with Clean Architecture"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
    "beautifulsoup4>=4.12.0",
    "playwright>=1.55.0",
    "facebook-business>=23.0.2",
    "apify-client>=2.1.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
]
